Add migration from carrierwave
This includes a lot of try-and-error, but not tests. Sorry.

It handles carrierwave-uploaders that reference one upload. The models
can be STI-classes. Uploads to both the base-class and the subclasses
are migrates.

During migration, the models are extended to have a has_one_attached.
The name is the old uploader prefixed with carrierwave.

A future extension might offer an automated migration to a not prefixed
name. The prefix is needed to distinguish between the old and new
uploader. It was easier to create a prefixed new version than to
retroactively prefix the old one. Since the new name is stored in the
DB, this could be done with in a migration later.